requirements.txt
amqp==5.1.1
async-timeout==4.0.2
billiard==4.1.0
blinker==1.6.2
cachelib==0.9.0
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
fasteners==0.18
Flask==2.2.5
Flask-Caching==2.0.2
Flask-MySQLdb==1.0.1
flask-paginate==2022.1.8
Flask-Pagination==0.0.1
Flask-Session==0.4.1
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.1
MarkupSafe==2.1.2
mysqlclient==2.1.1
numpy==1.25.1
pandas==2.0.3
prompt-toolkit==3.0.39
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyuwsgi==2.0.21
redis==4.6.0
requests==2.30.0
six==1.16.0
steamsignin==1.1.1
typing-extensions==4.7.1
tzdata==2023.3
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.3
xmltodict==0.13.0
zipp==3.15.0
